Use u64 for big numbers

+++ aoc14/src/data.rs
@@ -33,14 +33,14 @@ pub fn rules_to_hashmap(rules: Vec<Rule>) -> HashMap<String, char> {
 pub struct Polymer {
     pub sequence: String,
     pub rules: HashMap<String, char>,
-    pub tally: HashMap<String, u32>,
+    pub tally: HashMap<String, u64>,
 }
 
 impl Polymer {
     pub fn new(sequence: String, rules: Vec<Rule>) -> Self {
         let rules = rules_to_hashmap(rules);
 
-        let mut tally: HashMap<String, u32> = HashMap::new();
+        let mut tally: HashMap<String, u64> = HashMap::new();
         for index in 0..sequence.len() - 1 {
             let substr = &sequence[index..index + 2].to_string();
             let count = tally.entry(substr.to_string()).or_insert(0);
@@ -81,7 +81,7 @@ impl Polymer {
         }
     }
 
-    pub fn run_steps(self: &mut Self, num_steps: u32) {
+    pub fn run_steps(self: &mut Self, num_steps: u64) {
         for _ in 0..num_steps {
             self.step();
         }
